+++ public/app/features/alerting/unified/components/rule-editor/AnnotationsStep.tsx
@@ -11,6 +11,7 @@ import { t, Trans } from 'app/core/internationalization';
 import { DashboardModel } from '../../../../dashboard/state';
 import { RuleFormValues } from '../../types/rule-form';
 import { Annotation, annotationLabels } from '../../utils/constants';
+import { isGrafanaManagedRuleByType } from '../../utils/rules';
 
 import AnnotationHeaderField from './AnnotationHeaderField';
 import DashboardAnnotationField from './DashboardAnnotationField';
@@ -31,6 +32,7 @@ const AnnotationsStep = () => {
     setValue,
   } = useFormContext<RuleFormValues>();
   const annotations = watch('annotations');
+  const type = watch('type');
 
   const { fields, append, remove } = useFieldArray({ control, name: 'annotations' });
 
@@ -104,10 +106,12 @@ const AnnotationsStep = () => {
       </Stack>
     );
   }
+  // when using Grafana managed rules, the annotations step is the 6th step, as we have an additional step for the configure labels and notifications
+  const step = isGrafanaManagedRuleByType(type) ? 6 : 5;
 
   return (
     <RuleEditorSection
-      stepNo={5}
+      stepNo={step}
       title={t('alerting.annotations.title', 'Configure notification message')}
       description={getAnnotationsSectionDescription()}
       fullWidth
