+++ public/app/features/alerting/unified/components/rule-editor/alert-rule-form/ModifyExportRuleForm.tsx
@@ -24,7 +24,8 @@ import { GrafanaExportDrawer } from '../../export/GrafanaExportDrawer';
 import { ExportFormats, allGrafanaExportProviders } from '../../export/providers';
 import { AlertRuleNameAndMetric } from '../AlertRuleNameInput';
 import AnnotationsStep from '../AnnotationsStep';
-import { GrafanaEvaluationBehavior } from '../GrafanaEvaluationBehavior';
+import { GrafanaEvaluationBehaviorStep } from '../GrafanaEvaluationBehavior';
+import { GrafanaFolderAndLabelsStep } from '../GrafanaFolderAndLabelsStep';
 import { NotificationsStep } from '../NotificationsStep';
 import { QueryAndExpressionsStep } from '../query-and-alert-condition/QueryAndExpressionsStep';
 
@@ -90,15 +91,15 @@ export function ModifyExportRuleForm({ ruleForm, alertUid }: ModifyExportRuleFor
               {/* Step 2 */}
               <QueryAndExpressionsStep editingExistingRule={existing} onDataChange={checkAlertCondition} />
               {/* Step 3-4-5 */}
+              <GrafanaFolderAndLabelsStep />
 
-              <GrafanaEvaluationBehavior
+              {/* Step 4 & 5 */}
+              <GrafanaEvaluationBehaviorStep
                 evaluateEvery={evaluateEvery}
                 setEvaluateEvery={setEvaluateEvery}
                 existing={Boolean(existing)}
                 enableProvisionedGroups={true}
               />
-
-              {/* Step 4 & 5 */}
               {/* Notifications step*/}
               <NotificationsStep alertUid={alertUid} />
               {/* Annotations only for cloud and Grafana */}
