+++ public/app/features/alerting/unified/components/rule-editor/labels/LabelsEditorModal.tsx
@@ -1,5 +1,7 @@
 import { Modal } from '@grafana/ui';
 
+import { KBObjectArray } from '../../../types/rule-form';
+
 import { LabelsSubForm } from './LabelsField';
 
 export interface LabelsEditorModalProps {
@@ -8,12 +10,7 @@ export interface LabelsEditorModalProps {
     key: string;
     value: string;
   }>;
-  onClose: (
-    labelsToUodate?: Array<{
-      key: string;
-      value: string;
-    }>
-  ) => void;
+  onClose: (labelsToUodate?: KBObjectArray) => void;
   dataSourceName: string;
 }
 export function LabelsEditorModal({ isOpen, onClose, dataSourceName, initialLabels }: LabelsEditorModalProps) {
