+++ public/locales/pseudo-LOCALE/grafana.json
@@ -33,41 +33,11 @@
       }
     }
   },
-  "alert-rule-form": {
-    "evaluation-behaviour": {
-      "info-help": {
-        "text": "Đęƒįŉę ŧĥę äľęřŧ þęĥävįőř ŵĥęŉ ŧĥę ęväľūäŧįőŉ ƒäįľş őř ŧĥę qūęřy řęŧūřŉş ŉő đäŧä."
-      },
-      "pending-period": "Pęŉđįŉģ pęřįőđ"
-    },
-    "evaluation-behaviour-description1": "Ēväľūäŧįőŉ ģřőūpş äřę čőŉŧäįŉęřş ƒőř ęväľūäŧįŉģ äľęřŧ äŉđ řęčőřđįŉģ řūľęş.",
-    "evaluation-behaviour-description2": "Åŉ ęväľūäŧįőŉ ģřőūp đęƒįŉęş äŉ ęväľūäŧįőŉ įŉŧęřväľ - ĥőŵ őƒŧęŉ ä řūľę įş ęväľūäŧęđ. Åľęřŧ řūľęş ŵįŧĥįŉ ŧĥę şämę ęväľūäŧįőŉ ģřőūp äřę ęväľūäŧęđ ővęř ŧĥę şämę ęväľūäŧįőŉ įŉŧęřväľ.",
-    "evaluation-behaviour-description3": "Pęŉđįŉģ pęřįőđ şpęčįƒįęş ĥőŵ ľőŉģ ŧĥę ŧĥřęşĥőľđ čőŉđįŧįőŉ mūşŧ þę męŧ þęƒőřę ŧĥę äľęřŧ şŧäřŧş ƒįřįŉģ. Ŧĥįş őpŧįőŉ ĥęľpş přęvęŉŧ äľęřŧş ƒřőm þęįŉģ ŧřįģģęřęđ þy ŧęmpőřäřy įşşūęş.",
-    "evaluation-behaviour-for": {
-      "error-parsing": "Fäįľęđ ŧő päřşę đūřäŧįőŉ",
-      "validation": "Pęŉđįŉģ pęřįőđ mūşŧ þę ģřęäŧęř ŧĥäŉ őř ęqūäľ ŧő ŧĥę ęväľūäŧįőŉ įŉŧęřväľ."
-    },
-    "evaluation-behaviour-group": {
-      "text": "Åľľ řūľęş įŉ ŧĥę şęľęčŧęđ ģřőūp äřę ęväľūäŧęđ ęvęřy {{evaluateEvery}}."
-    },
-    "pause": {
-      "alerting": "Ŧūřŉ őŉ ŧő päūşę ęväľūäŧįőŉ ƒőř ŧĥįş äľęřŧ řūľę.",
-      "label": "Päūşę ęväľūäŧįőŉ",
-      "recording": "Ŧūřŉ őŉ ŧő päūşę ęväľūäŧįőŉ ƒőř ŧĥįş řęčőřđįŉģ řūľę."
-    }
-  },
   "alerting": {
     "alert-recording-rule-form": {
       "evaluation-behaviour": {
         "description": {
           "text": "Đęƒįŉę ĥőŵ ŧĥę řęčőřđįŉģ řūľę įş ęväľūäŧęđ."
-        }
-      }
-    },
-    "alert-rule-form": {
-      "evaluation-behaviour": {
-        "description": {
-          "text": "Đęƒįŉę ĥőŵ ŧĥę äľęřŧ řūľę įş ęväľūäŧęđ."
         }
       }
     },
@@ -177,10 +147,6 @@
             "alerting": "Cřęäŧę ä ŉęŵ ęväľūäŧįőŉ ģřőūp ŧő ūşę ƒőř ŧĥįş äľęřŧ řūľę.",
             "recording": "Cřęäŧę ä ŉęŵ ęväľūäŧįőŉ ģřőūp ŧő ūşę ƒőř ŧĥįş řęčőřđįŉģ řūľę."
           }
-        },
-        "text": {
-          "alerting": "Đęƒįŉę ĥőŵ őƒŧęŉ ŧĥę äľęřŧ řūľę įş ęväľūäŧęđ.",
-          "recording": "Đęƒįŉę ĥőŵ őƒŧęŉ ŧĥę řęčőřđįŉģ řūľę įş ęväľūäŧęđ."
         }
       }
     },
@@ -265,6 +231,60 @@
       },
       "preview": "Přęvįęŵ",
       "previewCondition": "Přęvįęŵ äľęřŧ řūľę čőŉđįŧįőŉ"
+    },
+    "rule-form": {
+      "evaluation": {
+        "evaluation-group-and-interval": "Ēväľūäŧįőŉ ģřőūp äŉđ įŉŧęřväľ",
+        "group": {
+          "cancel": "Cäŉčęľ",
+          "create": "Cřęäŧę",
+          "interval": "Ēväľūäŧįőŉ įŉŧęřväľ"
+        },
+        "group-name": "Ēväľūäŧįőŉ ģřőūp ŉämę",
+        "group-text": "Åľľ řūľęş įŉ ŧĥę şęľęčŧęđ ģřőūp äřę ęväľūäŧęđ ęvęřy {{evaluateEvery}}.",
+        "new-group": "Ńęŵ ęväľūäŧįőŉ ģřőūp",
+        "pause": {
+          "alerting": "Ŧūřŉ őŉ ŧő päūşę ęväľūäŧįőŉ ƒőř ŧĥįş äľęřŧ řūľę.",
+          "recording": "Ŧūřŉ őŉ ŧő päūşę ęväľūäŧįőŉ ƒőř ŧĥįş řęčőřđįŉģ řūľę."
+        },
+        "select-folder-before": "Ŝęľęčŧ ä ƒőľđęř þęƒőřę şęŧŧįŉģ ęväľūäŧįőŉ ģřőūp äŉđ įŉŧęřväľ"
+      },
+      "evaluation-behaviour": {
+        "description": {
+          "text": "Đęƒįŉę ĥőŵ ŧĥę äľęřŧ řūľę įş ęväľūäŧęđ."
+        },
+        "info-help": {
+          "text": "Đęƒįŉę ŧĥę äľęřŧ þęĥävįőř ŵĥęŉ ŧĥę ęväľūäŧįőŉ ƒäįľş őř ŧĥę qūęřy řęŧūřŉş ŉő đäŧä."
+        },
+        "pending-period": "Pęŉđįŉģ pęřįőđ"
+      },
+      "evaluation-behaviour-description1": "Ēväľūäŧįőŉ ģřőūpş äřę čőŉŧäįŉęřş ƒőř ęväľūäŧįŉģ äľęřŧ äŉđ řęčőřđįŉģ řūľęş.",
+      "evaluation-behaviour-description2": "Åŉ ęväľūäŧįőŉ ģřőūp đęƒįŉęş äŉ ęväľūäŧįőŉ įŉŧęřväľ - ĥőŵ őƒŧęŉ ä řūľę įş ęväľūäŧęđ. Åľęřŧ řūľęş ŵįŧĥįŉ ŧĥę şämę ęväľūäŧįőŉ ģřőūp äřę ęväľūäŧęđ ővęř ŧĥę şämę ęväľūäŧįőŉ įŉŧęřväľ.",
+      "evaluation-behaviour-description3": "Pęŉđįŉģ pęřįőđ şpęčįƒįęş ĥőŵ ľőŉģ ŧĥę ŧĥřęşĥőľđ čőŉđįŧįőŉ mūşŧ þę męŧ þęƒőřę ŧĥę äľęřŧ şŧäřŧş ƒįřįŉģ. Ŧĥįş őpŧįőŉ ĥęľpş přęvęŉŧ äľęřŧş ƒřőm þęįŉģ ŧřįģģęřęđ þy ŧęmpőřäřy įşşūęş.",
+      "evaluation-behaviour-for": {
+        "error-parsing": "Fäįľęđ ŧő päřşę đūřäŧįőŉ",
+        "validation": "Pęŉđįŉģ pęřįőđ mūşŧ þę ģřęäŧęř ŧĥäŉ őř ęqūäľ ŧő ŧĥę ęväľūäŧįőŉ įŉŧęřväľ."
+      },
+      "folder": {
+        "cancel": "Cäŉčęľ",
+        "create": "Cřęäŧę",
+        "create-folder": "Cřęäŧę ä ŉęŵ ƒőľđęř ŧő şŧőřę yőūř äľęřŧ řūľę įŉ.",
+        "creating-new-folder": "Cřęäŧįŉģ ŉęŵ ƒőľđęř",
+        "label": "Főľđęř",
+        "name": "Főľđęř ŉämę",
+        "new-folder": "Ńęŵ ƒőľđęř",
+        "new-folder-or": "őř"
+      },
+      "folder-and-labels": "Øřģäŉįžę yőūř äľęřŧ řūľę ŵįŧĥ ä ƒőľđęř äŉđ şęŧ őƒ ľäþęľş.",
+      "folders": {
+        "help-info": "Főľđęřş äřę ūşęđ ƒőř şŧőřįŉģ äľęřŧ řūľęş. Ÿőū čäŉ ęχŧęŉđ ŧĥę äččęşş přővįđęđ þy ä řőľę ŧő äľęřŧ řūľęş äŉđ äşşįģŉ pęřmįşşįőŉş ŧő įŉđįvįđūäľ ƒőľđęřş."
+      },
+      "labels": {
+        "help-info": "Ŀäþęľş äřę ūşęđ ŧő đįƒƒęřęŉŧįäŧę äŉ äľęřŧ ƒřőm äľľ őŧĥęř äľęřŧş.Ÿőū čäŉ ūşę ŧĥęm ƒőř şęäřčĥįŉģ, şįľęŉčįŉģ, äŉđ řőūŧįŉģ ŉőŧįƒįčäŧįőŉş."
+      },
+      "pause": {
+        "label": "Päūşę ęväľūäŧįőŉ"
+      }
     },
     "rule-groups": {
       "delete": {
